LPD-47204 Refactor code
