+++ source-formatter.properties
@@ -546,7 +546,23 @@
         static,\
         web-experience
 
-    source.check.CSPComplianceCheck.ignoredFTLTagPrefixes=\
+    source.check.CSPIllegalTagsCheck.ftlAndVmIllegalTagNames=\
+        embed,\
+        object
+
+    source.check.CSPIllegalTagsCheck.jspIllegalTagNames=\
+        embed,\
+        link,\
+        object,\
+        script,\
+        style
+
+    source.check.CSPIllegalTagsCheck.replacedTagNames=\
+        link,\
+        script,\
+        style
+
+    source.check.CSPTagIllegalAttributesCheck.ignoredFTLTagPrefixes=\
         <@clay.,\
         <@clay[,\
         <@liferay_aui.,\
@@ -556,13 +572,13 @@
         <@liferay_ui.,\
         <@liferay_ui[
 
-    source.check.CSPComplianceCheck.ignoredJSPTagPrefixes=\
+    source.check.CSPTagIllegalAttributesCheck.ignoredJSPTagPrefixes=\
         <aui:,\
         <clay:,\
         <liferay-frontend:,\
         <liferay-ui:
 
-    source.check.CSPComplianceCheck.illegalAttributeNames=\
+    source.check.CSPTagIllegalAttributesCheck.illegalAttributeNames=\
         onafterprint,\
         onbeforematch,\
         onbeforeprint,\
@@ -609,10 +625,8 @@
         onunload,\
         onvisibilitychange
 
-    source.check.CSPComplianceCheck.illegalTagNamesData=\
-        embed,\
-        link:rel="stylesheet",\
-        object,\
+    source.check.CSPTagMissingAttributesCheck.nonceAttributeTagNames=\
+        link,\
         script,\
         style
